[change]作成済みstoriesファイルのフォルダ分け（/common）#856
