+++ view/next-project/src/components/common/index.ts
@@ -13,6 +13,7 @@ export { default as OutlinePrimaryButton } from './OutlinePrimaryButton';
 export { default as OpenModalButton } from './OpenModalButton';
 export { default as PrimaryButton } from './PrimaryButton';
 export { default as PullDown } from './PullDown';
+export { default as PulldownButton } from './PulldownButton';
 export { default as Radio } from './Radio';
 export { default as RedButton } from './RedButton';
 export { default as RegistButton } from './RegistButton';
